requirements.txt
#Required Python packages
streamlit
pandas
scikit-learn
joblib
pymongo
dnspython
streamlit-extras
plotly
python-whois
